added home button to cam header & moved style sheets to be in components/removed styles folder
