+++ App.js
@@ -1,5 +1,4 @@
 import React from "react";
-// import { StyleSheet, Text, View, Button } from "react-native";
 import { NavigationContainer } from "@react-navigation/native";
 import { createStackNavigator } from "@react-navigation/stack";
 import CameraScreen from "./screens/CameraScreen";
