+++ screens/CameraScreen.js
@@ -7,14 +7,13 @@ import Settings from "../components/modals/SettingsModal";
 import LanguageModal from "../components/modals/LanguageModal";
 import Error from "../components/modals/ErrorModal";
 import Confirmation from "../components/modals/ConfirmationModal";
-import styles from "../styles/CameraStyle";
 import EditText from "../components/modals/EditTextModal";
 import LoadingWheel from "../components/LoadingWheel";
 import FlipButton from "../components/FlipButton";
 import FlashButton from "../components/FlashButton";
 import { Ionicons } from "@expo/vector-icons";
 import Languages from "../languages";
-import Header from "../Header";
+import Header from "../components/headers/Header";
 
 //Choosing a functional component gives us access to useState hook
 const CameraScreen = ({
@@ -104,8 +103,8 @@ const CameraScreen = ({
     return <Text>No access to camera</Text>;
   }
   return (
-    <View style={styles.container}>
-      <Header title="Point & Translate" />
+    <View style={styles.screen}>
+      <Header title="Point & Translate" navigation={navigation}/>
       <Camera
         ref={(ref) => setCamera(ref)}
         style={styles.camera}
@@ -140,7 +139,6 @@ const CameraScreen = ({
                 color={"white"}
                 style={styles.selectText}
               />
-              {/* <Text style={styles.selectText}>Settings</Text> */}
             </TouchableOpacity>
           </View>
           <View style={styles.cameraControlContainer}>
@@ -184,7 +182,6 @@ const CameraScreen = ({
 };
 
 const mapStateToProps = (state) => {
-  // console.log('state', state)
   return {
     orgText: state.source.detectedText,
     error: state.source.error,
@@ -199,5 +196,92 @@ const mapDispatchToProps = (dispatch) => {
     getText: (pic, ocrType) => dispatch(getText(pic, ocrType)),
   };
 };
+
+const styles = StyleSheet.create({
+  screen: {
+    flex: 1,
+  },
+  camera: {
+    flex: 1,
+  },
+  topButtons: {
+    flex: 1,
+    padding: 20,
+    height: 500,
+    flexDirection: "row",
+    justifyContent: "space-between",
+  },
+  buttonContainer: {
+    flex: 1,
+    backgroundColor: "transparent",
+    flexDirection: "column",
+    justifyContent: "space-between",
+    margin: 10,
+  },
+  button: {
+    flex: 0.1,
+    alignSelf: "flex-end",
+    alignItems: "center",
+  },
+  flipButton: {
+    alignSelf: "flex-start",
+  },
+
+  selectText: {
+    textAlign: "center",
+    padding: 5,
+    borderRadius: 10,
+    marginTop: 1,
+    marginBottom: 3,
+    color: "white",
+    fontSize: 20,
+  },
+  languageButton: {
+    flexDirection: "row",
+    flex: 0.6,
+    alignSelf: "flex-start",
+    fontSize: 18,
+    height: 40,
+    width: 150,
+    backgroundColor: "rgba(0,0,0,0.5)",
+    borderRadius: 10,
+    alignItems: "center",
+    justifyContent: "center",
+  },
+  langButton: {
+    alignSelf: "flex-start",
+    fontSize: 18,
+    backgroundColor: "rgba(0,0,0,0.5)",
+    height: 35,
+    borderRadius: 10,
+    width: 50,
+  },
+  shutterButton: {
+    paddingVertical: 10,
+    paddingHorizontal: 10,
+    padding: 40,
+    alignSelf: "center",
+    // backgroundColor: "#D90E18",
+    // backgroundColor: "#009FB8",
+    backgroundColor: "#FC9E9C",
+    // borderColor: "#006575",
+    borderColor: "#FB7573",
+    borderRadius: 50,
+    borderWidth: 8,
+    width: 80,
+    height: 80,
+    margin: 20,
+  },
+  flipButtonText: {
+    fontSize: 18,
+    color: "white",
+  },
+  cameraControlContainer: {
+    flex: 0.2,
+    flexDirection: "row",
+    justifyContent: "space-between",
+    alignContent: "center",
+  },
+});
 
 export default connect(mapStateToProps, mapDispatchToProps)(CameraScreen);
